requirements.txt
tensorflow==2.19.0
opencv-python==4.12.0.88
pillow==11.3.0
numpy==1.26.0
matplotlib==3.10.5
scikit-learn==1.7.1
streamlit==1.48.0
huggingface-hub==0.18.1
